+++ packages/core/code/examples/11-code-block-dark-theme.js
@@ -1,0 +1,29 @@
+// @flow
+import React from 'react';
+import { AtlasKitThemeProvider } from '@atlaskit/theme';
+import { AkCodeBlock } from '../src';
+
+const exampleCodeBlock = `
+  class HelloMessage extends React.Component {
+    render() {
+      return (
+        <div>
+          Hello {this.props.name}
+        </div>
+      );
+    }
+  }
+
+  ReactDOM.render(
+    <HelloMessage name="Taylor" />,
+    mountNode
+  );
+`;
+
+export default function Component() {
+  return (
+    <AtlasKitThemeProvider mode="dark">
+      <AkCodeBlock language="java" text={exampleCodeBlock} />
+    </AtlasKitThemeProvider>
+  );
+}