+++ packages/core/datetime-picker/examples/00-basic.js
@@ -90,15 +90,14 @@ export default () => {
 
       <Label
         htmlFor="react-select-datepicker-2--input"
-        label="controlled (value, isOpen)"
+        label="controlled (value)"
       />
       <Controlled initialValue="2018-01-02">
-        {({ value, onValueChange, isOpen, onBlur }) => (
+        {({ value, onValueChange, onBlur }) => (
           <DatePicker
             id="datepicker-2"
             value={value}
             onChange={onValueChange}
-            isOpen={isOpen}
             onBlur={onBlur}
           />
         )}