+++ packages/editor/editor-core/src/plugins/collab-edit/actions.ts
@@ -14,18 +14,20 @@ import {
 export const handleInit = (initData: InitData, view: EditorView) => {
   const { doc, json } = initData;
   if (doc) {
-    const { state, state: { schema, tr } } = view;
+    const {
+      state,
+      state: { schema, tr },
+    } = view;
     const content = (doc.content || []).map(child =>
       schema.nodeFromJSON(child),
     );
 
     if (content.length) {
-      const newState = state.apply(
-        tr
-          .setMeta('addToHistory', false)
-          .replaceWith(0, state.doc.nodeSize - 2, content)
-          .scrollIntoView(),
-      );
+      tr.setMeta('addToHistory', false);
+      tr.replaceWith(0, state.doc.nodeSize - 2, content);
+      tr.setSelection(Selection.atStart(tr.doc));
+      tr.scrollIntoView();
+      const newState = state.apply(tr);
       view.updateState(newState);
     }
   } else if (json) {
@@ -37,7 +39,9 @@ export const handleConnection = (
   connectionData: ConnectionData,
   view: EditorView,
 ) => {
-  const { state: { tr } } = view;
+  const {
+    state: { tr },
+  } = view;
   view.dispatch(tr.setMeta('sessionId', connectionData));
 };
 
@@ -45,7 +49,9 @@ export const handlePresence = (
   presenceData: PresenceData,
   view: EditorView,
 ) => {
-  const { state: { tr } } = view;
+  const {
+    state: { tr },
+  } = view;
   view.dispatch(tr.setMeta('presence', presenceData));
 };
 
@@ -59,7 +65,10 @@ export const applyRemoteData = (remoteData: RemoteData, view: EditorView) => {
 };
 
 export const applyRemoteSteps = (json: any[], view: EditorView) => {
-  const { state, state: { schema } } = view;
+  const {
+    state,
+    state: { schema },
+  } = view;
   let { tr } = state;
 
   json.forEach(stepJson => {
@@ -77,7 +86,9 @@ export const handleTelePointer = (
   telepointerData: TelepointerData,
   view: EditorView,
 ) => {
-  const { state: { tr } } = view;
+  const {
+    state: { tr },
+  } = view;
   view.dispatch(tr.setMeta('telepointer', telepointerData));
 };