+++ packages/media/media-picker/src/service/newUploadServiceImpl.ts
@@ -5,7 +5,7 @@ import {
   MediaType,
   FileItem,
 } from '@atlaskit/media-core';
-import { MediaStore } from '@atlaskit/media-store';
+import { MediaStore, UploadController } from '@atlaskit/media-store';
 import { EventEmitter2 } from 'eventemitter2';
 import { defaultUploadParams } from '../domain/uploadParams';
 import { MediaFile, PublicMediaFile } from '../domain/file';
@@ -57,11 +57,16 @@ export class NewUploadServiceImpl implements UploadService {
       ...uploadParams,
     };
   }
+  // Used for testing
+  private createUploadController(): UploadController {
+    return new UploadController();
+  }
 
   addFiles(files: File[]): void {
     if (files.length === 0) {
       return;
     }
+
     const creationDate = Date.now();
     const cancellableFileUploads: CancellableFileUpload[] = files.map(file => ({
       mediaFile: {
@@ -90,17 +95,29 @@ export class NewUploadServiceImpl implements UploadService {
         name: file.name,
         mimeType: file.type,
       };
-      const { deferredFileId, cancel } = this.context.uploadFile(
-        uploadableFile,
-        {
-          onProgress: this.onFileProgress.bind(this, cancellableFileUpload),
-        },
-      );
-      cancellableFileUpload.cancel = cancel;
-      deferredFileId.then(
-        this.onFileSuccess.bind(this, cancellableFileUpload),
-        this.onFileError.bind(this, mediaFile, 'upload_fail'),
-      );
+      const controller = this.createUploadController();
+      const subscrition = this.context
+        .uploadFile(uploadableFile, controller)
+        .subscribe({
+          next: state => {
+            if (state.status === 'uploading') {
+              this.onFileProgress(cancellableFileUpload, state.progress);
+            }
+
+            if (state.status === 'processing') {
+              subscrition.unsubscribe();
+              this.onFileSuccess(cancellableFileUpload, state.id);
+            }
+          },
+          error: error => {
+            this.onFileError(mediaFile, 'upload_fail', error);
+          },
+        });
+
+      cancellableFileUpload.cancel = () => {
+        // we can't do "cancellableFileUpload.cancel = controller.abort" because will change the "this" context
+        controller.abort();
+      };
     });
   }